% Generated by roxygen2 (4.0.0): do not edit by hand
\name{age_split}
\alias{age_split}
\title{Age split event history data}
\usage{
age_split(dat, start = "start", end = "end", id = "id", event = "event",
  split_by = 5)
}
\arguments{
  \item{dat}{data.frame containing event history data}

  \item{start}{= "start" Name of start column}

  \item{end}{= "end" Name of end column}

  \item{id}{= "id" Name of id column}

  \item{event}{= "event" Name of event column}

  \item{split_by}{= 5 Age group size}
}
\description{
Split event history data into a sequence of age groups. A row is split if it the range
  between start and end stretches over a age group. Each row is assigned a new start
  and end point, a group and a duration within that group.
}

