% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_zph.R
\name{as.data.frame.cox.zph}
\alias{as.data.frame.cox.zph}
\title{cox.zph as data.frame}
\usage{
\method{as.data.frame}{cox.zph}(x, val = NA)
}
\arguments{
\item{x}{a cox.zph object}

\item{val}{index of variables to extract}
}
\description{
Extract scaled Schoenfeld residuals by variable for a cox.zph object
}
\examples{
library(survival)
fit <- coxph(Surv(time, status) ~ age + sex + disease, data = kidney)
fit_zph <- cox.zph(fit, transform = "identity")
zph_df <- as.data.frame(fit_zph)

}

