% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_coxme.R
\name{augment.coxme}
\alias{augment.coxme}
\title{Augment coxme}
\usage{
\method{augment}{coxme}(x, data = stats::model.frame(x), newdata = NULL,
  type.predict = "lp", ...)
}
\arguments{
\item{x}{coxme object}

\item{data}{original data for augment}

\item{newdata}{new data on which to do predictions}

\item{type.predict}{type of prediction value linear predictor or risk}

\item{...}{Extra arguments, not used}
}
\description{
Augment coxme object
}
\examples{
library(broom)
library(coxme)
fit <- coxme(Surv(y, uncens) ~ trt + (1|center), eortc)
augment(fit, type.predict = "risk")

}
