% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cox_table.R
\name{cox_table}
\alias{cox_table}
\title{Make a Results Table From Cox Regression}
\usage{
cox_table(cox)
}
\arguments{
\item{cox}{A cox.ph regression}
}
\description{
Make a Results Table From Cox Regression
}

