% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cox_table.R
\name{coxph_to_long}
\alias{coxph_to_long}
\title{Transform coxph objects to long format table}
\usage{
coxph_to_long(..., cilevel = 0.95)
}
\arguments{
\item{...}{coxph objects or list of objects}

\item{cilevel}{confidence interval}
}
\description{
For nested models
}
\examples{
library(survival)
data('cancer', package = 'survival')
cancer$sex <- factor(cancer$sex, labels = c('male', 'female'))
model1 <- coxph(Surv(time, status) ~ age + sex  + ph.ecog + strata(inst), data=cancer)
model2 <- coxph(Surv(time, status) ~ age + strata(inst), data=cancer)
coxph_to_long(model1, model2)
}
\author{
Johan Junkka \email{johan.junkka@gmail.com}
}

