% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_coxme.R
\name{glance.coxme}
\alias{glance.coxme}
\title{Glance coxme}
\usage{
\method{glance}{coxme}(x, ...)
}
\arguments{
\item{x}{coxme object}

\item{...}{other params}
}
\description{
Glance method for coxme objects
}
\examples{
library(broom)
library(coxme)
fit <- coxme(Surv(y, uncens) ~ trt + (1|center), eortc)
glance(fit)

}

