% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/broom_coxreg.R
\name{glance.coxreg}
\alias{glance.coxreg}
\title{Glance coxreg}
\usage{
\method{glance}{coxreg}(x, ...)
}
\arguments{
\item{x}{coxme object}

\item{...}{other params}
}
\description{
Glance method for coxreg objects
}

