% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_pred.R
\name{rr_pred.coxme}
\alias{rr_pred.coxme}
\title{Predict relative risk}
\usage{
\method{rr_pred}{coxme}(model, newdata = NULL, center, conf.level = 0.95)
}
\arguments{
\item{model}{coxme object}

\item{newdata}{data.frame}

\item{center}{center variable}

\item{conf.level}{The confidence level to use for the confidence interval.}
}
\description{
predict centered relative risk from  coxme object
}
\examples{

library(coxme)
fit <- coxme(Surv(time, status) ~ age + ph.ecog + (1 | inst), data = lung)
newdata <- expand.grid(age = 50, ph.ecog = 0:3, inst = 11)
rr_pred(fit, newdata, center = c(ph.ecog = 0))
}
