% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_pred.R
\name{rr_pred}
\alias{rr_pred}
\title{Predict relative risk}
\usage{
rr_pred(x, ...)
}
\arguments{
\item{x}{coxph object}

\item{...}{others}
}
\description{
predict centered relative risk from coxph and coxme objects
}
