% Generated by roxygen2 (4.0.0): do not edit by hand
\name{split_rows}
\alias{split_rows}
\title{Split Rows}
\usage{
split_rows(df, splitAt, id = "id", start = "start", end = "end")
}
\arguments{
  \item{df}{A data.frame}

  \item{splitAt}{A number specifing the point within range
  to split at}

  \item{id}{The name of id column, default is "id"}

  \item{start}{The name of start of range column, default
  is "start"}

  \item{end}{The name of end of range column, default is
  "end"}
}
\value{
A data.frame with new start and end range, group, start_type and end_type
}
\description{
Split rows in data.frame if split point falls between start and end of range in data.frame
}
\author{
Johan Junkka
}

