% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_coxme.R
\name{tidy.coxme}
\alias{tidy.coxme}
\title{Tidy coxme}
\usage{
\method{tidy}{coxme}(x, exponentiate = FALSE, ...)
}
\arguments{
\item{x}{coxme object}

\item{exponentiate}{boolean}

\item{...}{other params}
}
\description{
Tidy method for coxme objects
}
\examples{
library(broom)
library(coxme)
fit <- coxme(Surv(y, uncens) ~ trt + (1|center), eortc)
tidy(fit)

}

