% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/broom_coxreg.R
\name{tidy.coxreg}
\alias{tidy.coxreg}
\title{Tidy coxreg}
\usage{
\method{tidy}{coxreg}(x, exponentiate = TRUE, ...)
}
\arguments{
\item{x}{coxreg object}

\item{exponentiate}{boolean}

\item{...}{other params}
}
\description{
Tidy method for coxme objects
}
\examples{
require(eha)
require(broom)
data(oldmort)
fit <- coxreg(Surv(enter, exit, event) ~ civ + sex + birthdate, data = oldmort)
tidy(fit)
}

