% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coxph_to_long}
\alias{coxph_to_long}
\title{Transform coxph objects to long format table}
\usage{
coxph_to_long(..., cilevel = 0.95)
}
\arguments{
\item{...}{coxph objects or list of objects}

\item{cilevel}{confidence interval}
}
\description{
For nested models
}
\author{
Johan Junkka \email{johan.junkka@gmail.com}
}

