% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_zph.R
\name{gg_zph}
\alias{gg_zph}
\title{ggplot cox.zph}
\usage{
gg_zph(x, val = NA, log = FALSE, alpha = 0.2)
}
\arguments{
\item{x}{a cox.zph object}

\item{val}{index of variables to plot}

\item{log}{logical, if TRUE log transform time}

\item{alpha}{geom_point alpha value}
}
\description{
A ggplot2 implementation of plotting scaled Schoenfeld residuals
  from a cox.zph objects
}
\examples{
library(survival)
fit <- coxph(Surv(time, status) ~ age + sex + disease, data = kidney)
fit_zph <- cox.zph(fit, transform = "identity")
gg_zph(fit_zph)
gg_zph(fit_zph, log = TRUE)

}

