% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ggsurv}
\alias{ggsurv}
\title{Survfit for ggplot2}
\usage{
ggsurv(x)
}
\arguments{
  \item{x}{a coxph or survfit object}
}
\value{
data.frame with time, strata, std_err, upper, lower, n_risk, n_event, n_centor,
  and surv. Optionaly if object is coxph, it returns cumhaz and if the object is
stratified, strata.
}
\description{
ggsurv creates a data.frame of a coxph or a survfit object for
  visualization with ggplot2
}
\examples{
\dontrun{
surv_object <- coxph(Surv(time, status) ~ strata(x), data = aml)
ggplot(ggsurv(surv_object), aes(time, surv, color=strata)) + geom_step()
}
}
\author{
Johan Junkka
}

