% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{merge_reg}
\alias{merge_reg}
\title{Merge Cox Regessions to One Table}
\usage{
merge_reg(full, list_res)
}
\arguments{
\item{full}{A model with all relavent variables or cector
f covariate names}

\item{list_res}{a list with all regressions to be merged}
}
\description{
Merge Cox Regessions to One Table
}

