% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_coxme.R
\name{predict_coxme}
\alias{predict_coxme}
\title{Predict coxme}
\usage{
predict_coxme(object, newdata = NULL, type = c("lp", "risk"),
  strata_ref = TRUE)
}
\arguments{
\item{object}{coxme object}

\item{newdata}{data.frame new data set}

\item{type}{type of prediction, linear predictor ("lp") or relative risk ("risk")}

\item{strata_ref}{logical, use strata as reference}
}
\description{
Prediction method for coxme objects
}
\examples{
library(coxme)
fit <- coxme(Surv(y, uncens) ~ trt + (1|center), eortc)
predict_coxme(fit)

}
