% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/broom_coxme.R
\name{tidy.coxme}
\alias{tidy.coxme}
\title{Tidy coxme}
\usage{
\method{tidy}{coxme}(x, exponentiate = TRUE, ...)
}
\arguments{
\item{x}{coxme object}

\item{exponentiate}{boolean}

\item{...}{other params}
}
\description{
Tidy method for coxme objects
}

