% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_coxme.R
\name{tidy.coxme}
\alias{tidy.coxme}
\title{Tidy coxme}
\usage{
\method{tidy}{coxme}(x, exponentiate = FALSE, conf.int = 0.95, ...)
}
\arguments{
\item{x}{coxme object}

\item{exponentiate}{whether to report the estimate and 
confidence intervals on an exponential scale}

\item{conf.int}{confidence level to be used for CI}

\item{...}{extra arguments, not used}
}
\description{
Tidy method for coxme objects
}
\examples{
library(broom)
library(coxme)
fit <- coxme(Surv(y, uncens) ~ trt + (1|center), eortc)
tidy(fit)

}

